Replace StM m a ~ a in the type signatures with Forall (Pure m) (#12)
This change is needed to simplify client code.
Suppose we have a function which spawns two
threads using `async` with different return types
`a` and `b` respectively. Then we need to add
`StM m a ~ a` and `StM m b ~ b` constraints to the
function, although what we really need was
`forall a. StM m a ~ a` which is polymorphic to the
type variable `a`.

`Forall` in constraints package is now re-exported
from the Safe module for convenience.