In .Safe module, use async's concurrently, race, and race_
The upstream versions of the functions are more efficient, being
implemented via MVars and not leveraging the full Async machinery. In
addition to getting a performance boost from this change, we also
guarantee that, in the case of deadlocks, the exceptions generated from
the .Safe module match those generated by Control.Concurrent.Async.

Note that this change does not make sense for the unsafe module in this
package.